Fixed title hierarchy problem.
The method for "underlining" everything after the main heading was
identical, which indicated that they were all of the same level. Now
subheadings are handled properly. This could be more useful if we decide
to used something like sphinx, which formats TOC entries based off of
how headings are "underlined".