Modified file hierarchy and names for better organization.
A lot of files were just mixed together in the main tutorial directory,
which was kind of messy. I updated the directions, so that each new main
section creates a brand new directory. Also, some of the file names were
changed to make things more clear. Lastly, a number of typos and
mistakes were fixed.