More minor clarifications targeted at new MCell users.
Note that many of the instructions reference names like "New Region" or "New Molecule", but the current code appears to be generating them with shorter names like "Region" and "Molecule". It might be good to have a way to note when these kinds of User Interface strings change so we can update our documentation.