Add local version of images
There is now a directory in the repo that stores all the images for the
tutorials. All the links in the ReST files have been updated to reflect
this. Previously, the images were only stored remotely to keep the repo
size small, but the new method is likely a better option given that
there are more collaborators now who will need to add/update the images.