Removed face 1 from top's ELEMENT_LIST
For some reason, "top" was shown with faces 1,8,9 (which would be 3 triangles instead of 2). The first 4 verticies (0-3) have negative Z values (bottom of the cube) while the second 4 verticies (4-7) have positive Z values (top of the box). So triangles for the top of the cube would have to consist of verticies with indexes between 4 and 7 inclusive. Only faces 8 and 9 from the example fit that requirement, so 1 was removed.