Updated "Getting Started" section.
I updated both the text and images for the "Getting Started" section to
be compatible with the the newest version of CellBlender (0.1.57).
Notable changes include the addition of "Project Settings", "Create
Reaction Output", and "Create Visualization Output".

I also made some smaller changes to the "Index", "Installs", and
"Software Needed" sections to bring version numbers up-to-date. Also,
some of the titles now mention CellBlender in addition to MCell.