Updated the "Visualize and Plot" section.
The video and some of the plotting information was commented out, since
it was out-of-date.

The **Visualize Simulation Results** panel has changed, so that section
was updated.

Also, it was pretty hard to see the glyphs because they were so small,
so I added a section where the user scales them up (along with some
pictures).