Initial version of the CellOrganelle12 tutorial
Pulled the text out of mcelltutorialtranscript.pdf (from Ben) and
watched the movie beng276_MCell_demo_3a.mov (also from Ben) to get
started. Spent a lot of time creating a GIF file to describe the
model that's being built (cellmodel.gif and cellmodel_s.gif).

I was able to run the demo all the way through, but the tutorial
is still just getting started.

This tutorial was started using:
  CellBlender Source ID = 51e3abf95f840f76e3af04c5bf4c2d5114122ebf