Updated tutorials to CellBlender 1.0 RC
This was mainly a matter of updating the "Note" message at the top of
every tutorial, which indicates the version of CellBlender and Blender
that were used to generate the tutorials.

The project setting images were updated, so that the SHA ID would be
the same as appears in 1.0 RC.

The default region name was also updated in one section to be compatible
with 1.0 RC.

The order of "Visualize and Plot" and "Examine the MDLs" were flipped.
Originally, changes needed to be made to the MDLs prior to
visualization, but now all those changes can be made within CellBlender,
so it seems less imperative to have users examine them immediately.