Made 1.0 and 1.0_RC directories
To prepare for the new tutorials, everything has been copied into two
separate directories, 1.0 and 1.0_RC. Note: these numbers refer to
CellBlender's version numbers and not the version numbers of the
tutorials themselves. Also, the naming is misleading, as one would
expect there to be minimal changes between a release candidate and the
final release, but the differences between the two are quite
substantial.

1.0 is just a plain copy of everything in 1.0_RC for now but will
eventually contain all the updates needed for CellBlender 1.0. The older
version will continue to be maintained, since some users may continue to
use it for some time to come. However, this hopefully will require few
if any changes in the future.