Updated the "Release Pattern" tutorial for CB 1.0
The previous tutorial relied on creating the model by hand-editing MDLs.
Now that release patterns can be created from within CB, the tutorial
has been udpated to reflect this. A plot of the reaction data with
overlays was also added so users can clearly see the effects that the
different release pattern properties have.