Update "Surf Classes & Surf Molec" tut for CB 1.0
Updated images and text for the "Surface Classes and Surface Molecules" so that
it is consistent with the changes in CB 1.0. Fixed some typos and broke up some
large blocks of text so it's easier to follow.

We should maybe consider merging this tutorial with "Introduction to Surface
Classes" or breaking that one up into two sections ("Surface Classes and Volume
Molecules" and "Surface Classes and Reactions").