Smaller sub-TOCs. Fixed typo. Updated intro.
I changed the max-depth of the table of contents for the "Advanced
Models" and "Advanced Features", so that they are a little easier to
parse (and hopefully less intimidating).

The introductory paragraph was updated to reflect some of the changes
that were made to the other tutorials (i.e. they are no longer dependent
on the "Getting Started" tutorial).