Added Jose and Devin's workshop demo as a tutorial
This tutorial is an adaptation of a demo that Jose and Devin gave at the
2014 MCell workshop. Some of the wording had to be changed to make it
appropriate for the format of a web-based tutorial. The initial
conversion of the tex file to rst was done with pandoc. Also, the way
that figures were done in LaTeX wasn't compatible with how they are
normally done in reStucturedText, so a custom sphinx extension was used
to allow it.