Streamlined the main tutorial page.
This commit is intended to address problems with certain sections being
hard to find. Links that were buried in paragraphs are now in a short
simple list. The introductory text is much more concise. Some of the
text that was on this page was moved to the quick start tutorial where
it fit better. A couple of links that were included in the "Advanced
Features" section were removed from there and added to this main table
of contents.