Added a Clean Slate tutorial, a Model Library, and a Single Molecule
The "Clean Slate" sets up a startup file with CellBlender and no
objects.

The Model Library is intended to be a series of models from simple to
complex showing various features of CellBlender and MCell.

The "Single Molecule Diffusion" is a tutorial to create and display a
single diffusing molecule.