Added syntax highlighting for MDL.
The MDL syntax highlighting uses a custom pygments lexer based off of
one of the built-in C-style lexers. It could be improved, but it works
reasonably well right now. MDL commands, string literals, and numbers
are all colored differently. Everything else is black.

The specific colors are determined by the pygment style, which was
changed from default to xcode. It might be worth finding/creating a
better one, since the current one uses green comments on a green
background.

To enable the custom syntax highlighting for a block of MDL, all one
needs to do is add ".. code-block:: mdl" before the block itself.