Add requirements.txt. Update README.
This is very similar to a commit that was recently made in QRG repo. As
such, I'm just going to copy-paste the message from that commit:

A pip requirements file was added to hopefully make it easier for people
to build the QRG and contribute to it. The README was updated to explain
how to use a Python virtual environment and a pip requirements file.