Removed clamp references and turned off counting plane generation.
The concentration clamp was left over from the single ended verion which
provided molecules via a clamp on the end at the origin. This change
removed the comments and unused code related to that clamp.

Also, the default parameter setting for "plot_front_crossings" was 1
which generated the counting planes. These are not used in the current
tutorial. This change sets it to zero. A future update of this tutorial
to provide flux measurements should set it back to 1.