Updated the Fick's model script to use a "seeds" parameter.
The previous version of this tutorial required the manual setting of the
number of seeds to run and then the manual updating of that number in
the plotting script. Because commit 85288bd4cb added start and end seeds
to the data model, these values can be stored in a parameter shared by
both the building and plotting scripts. This change implements that
approach creating a new paramater named "seeds" which sets the number of
seeds to run (1 to seeds).