Modified instructions to show selection of Geometry as needed.
There has been some discussion as to whether geometry is properly
included in a data model or not. The current development branch defines
the defaults for the optional "geometry" parameter as False for both
getting and replacing a data model. This update to the tutorials
reflects those changes by requiring the "geometry=True" setting in
various places.