This branch is a simplified version using Sphinx defaults.
The current master branch requires a number of new features to be
installed. This keeps the tutorials from building on some platforms.

This branch is intended to provide a simpler version that relies on
Sphinx defaults to increase the likelihood of building on a wider range
of platforms.

This particular commit commented out a number of features in "conf.py".