Updated the Dynamic Geometry section to plot dynamic volume via counts
This example uses MCell's concentration clamp to maintain a relatively
constant molecule density within a dynamic object. The count of the
molecules is roughly proportional to the volume of the object and can be
plotted to show the dynamic volume change.

Note that this example was constructed in the "default_style" branch and
should be merged into the master branch without disturbing the style
used in the master branch. Hopefully changes from the master branch can
also be merged into this "default_style" branch so they can be built
with the default Sphinx tools.