Removed dependency on "sphinx_rtd_theme" (puts next/prev at top).
The default_style branch is intended to build the "old style" pages that
include the "next" and "previous" links at the top of the page. The old
style pages also scale to use the full width of a browser rather than
being confined to a narrow column.

This change allowed building the default style version.

This commit fixes the following error when building in some
environments:

  Running Sphinx v1.2.3

  Exception occurred:
    File "conf.py", line 17, in <module>
      import sphinx_rtd_theme
  ImportError: No module named sphinx_rtd_theme