Added "htmldefault" option to the Makefile to bypass sphinx_rtd_theme
Rather than maintaining two separate branches, this commit adds a new
option to the Makefile to build using the default HTML style.

This is currently a very kludgy solution since it replaces the conf.py
file with conf_default_style.py before building (and swaps it back after
building). It was hoped that supplying command line arguments could
allow building with the same conf.py, but conf.py imports the
sphinx_rtd_theme which wouldn't be a command line option.

Further exploration might find a better approach, but this works for now.