Used -D option to pass arguments to conf.py to remove the kludgy copying
The previous version was a quick hack to build the tutorials with the
default (non-rtd) theme. This commit uses a much better approach.

The Makefile now calls sphinx-build with a -D option as follows:

  -D special_style=default

This ends up putting "special_style=default" into sys.argv when calling
conf.py. This special string is then used in conf.py to NOT include the
missing sphinx_rtd_theme module, and to make the other changes needed to
produce the default themed HTML.

Note that the default theme has a few features that make it handy for
viewing the tutorials:

  1. It generates HTML that can use the browser's full screen width
  2. It generates "previous next index" links at both bottom AND top.
     For some reason, the RTD theme only includes these links at the
     bottom of the page. This requires scrolling to the bottom of the
     page in order to advance from page to page through the tutorials.

To build HTML with the default theme simply use:

   make htmldefault