Change wording to "script insertion definition lines" and other changes
The export scripting interface consists of a list of scripts to be run
at various points in the exporting process. Each line of that list will
reference a single script to be run at a given point in the export.

In some of the previous descriptions, the elements in this list have
been referred to as "scripts" since they do represent scripts. But that
terminology can be confusing because they are not the scripts themselves
(which are, of course, text files).

To (hopefully) reduce any confusion, this update attempts to use the
terminology of "script insertion definition" to refer to the entries in
the list (which define where a script will be inserted). Since that's a
somewhat long name, it was shortened to "script definition" in places
where the context was hopfully clear.