Changed reaction rate in text from 5e3 to 1e8.
There was a conflict in the "6. Bimolecular Reactions" tutorial.

The text gives a reaction rate of 5e3. But the Panel image shows a
reaction rate of 1e8. If run with 5e3, there are no reactions. But
if run with 1e8, there's a nice curve from initial 1000 down to 250.

This commit changed the text to 1e8 to match the figure rather than
the other way around. This was done as a separate commit to call out
the change in case it should have been 5e3.