The Quick Tutorials section is an attempt to make easier tutorials
The current RST formats all end up generating a static document (whether
a web page, PDF, or other). This is very useful for documentation, but
there may be better ways to present tutorials.

This branch is an attempt to explore better ways to present tutorials as
well as easier ways to generate them. It is highly experimental.