Scroll the icon bar automatically when the next/prev slides are selected
Before this change, advancing the slides from the main window would
cause icons in the bar to be selected that were not visible. This made
it hard to tell the current step within the larger process. This change
scrolls the bar to attempt to show the currently selected icon. There
are some problems with this selection. It may require knowing the width
of each icon to do properly. But it works reasonably well for now.