Added a prototype tutorial format for web viewing.
This prototype uses HTML, CSS, and some Javascript to make a simple
tutorial that's intended to be easier to follow than our existing
tutorials. Each step in the process shows the entire Blender screen
to make it easier to follow along.

USE: Click on right side of main window to move to next slide.
     Click on left side of main window to move to previous slide.

This is a quick prototype that suffers from some known problems:
  - Resizing the browser can obscure the scrolling window
  - Advancing the slides doesn't exactly center the scrolled icon

These should be fixable if we use this format.