This version works reasonably well with a window size of about 1200x1200
This demonstration doesn't resize itself very well, so it's best viewed
in a browser window that's sized to about 1200x1200. At that size, the
scrolling images (along the bottom) should show about 4 previews. It
should also show a scroll bar just below those 4 images. The main window
should be clear enough to see, and the there should be sufficient room
for the text above and below the main Blender image.