Added Intro/Conclusion and fixed scroll bar positioning.
Each slide show should probably have an introduction and conclusion to
let the viewer know they've wrapped around in the slides. This commit
adds that for the SimpleReleaseTwoMolecules example.

The previous version didn't do a good job of setting the scroll position
based on the currently selected slide. This version does a better job on
that computation.