Added the Capsule test images to the dynamic geometry section.
The code to make these still exists as a separate file in the test_suite
subdirectory. That code is currently a separate Blender addon, and it
would be nice to eventually integrate it as an addon. For now, these
images (from November 18th, 2015) have been added to document the work.