Added description of "Average Coincident" option and supporting images
It's possible to specify geometries that don't "close" properly. An easy
example would be a square with 91 degree angles. The "Average Coincident"
option was added to "fix" these problems when they are small enough. The
added description and images (including a "fixed" double helix) show the
use (and potential misuse) of this option.