Added the Double Helix source code as an appendix to structured mols
The Double_Helix.py file was developed as a proof of concept challenge
for the Structured Molecules project. While it's not being released as
an official part of CellBlender, it is still useful to see how a fairly
complex molecule can be built using the current Structured Molecules
implementation. Note that this code builds a structure resembling the
GCAT structure of DNA, but it is NOT an accurate representation. It may
be possible to make changes to the parameters to make it more closely
resemble DNA itself, but that's not the current purpose of this code.
The code only serves as an example of what might be built with this
tool.