Added the original dialog-note.png file from the sphinx package.
The files used to build the HTML include a python installation with the
sphinx package installed in site-packages. That's where the built-in
images are stored. Since a subsequent commit will change one of the
images, this commit simply stores that original file. There may be a way
to put this in the more "proper" section of the repository, but at least
this commit documents what's being changed.