Changed the dialog-note.png image to be a small blue arrow.
The previous dialog-note.png icon was an orange ball with a white
exclamation point. This appears almost as an error or warning. But the
use of the "note" was not intended to be an error or a warning. For this
reason, another less alarming icon is being substituted. This commit
uses a small blue arrow, but other icons may be used.

The only tricky part about this commit is that this file is part of the
Python environment that runs sphinx. It's not clear how that Python
environment might be installed by other developers. Depending on how
this is done, the icon added in this commit might not actually be used
by sphinx when it builds the HTML or other output.