Removed the absolute address from the Fick's law video.
The PSC version had the full URL in the video link. This makes the page
less portable, so the URL was made relative. This should inherit the
protocol (http or https) from the parent. For some reason, refreshing
the page has sometimes shown a secure page and other times shown mixed
secure and insecure items. Refreshing the page seems to randomly show
either at the time of this commit.

Theoretically, most content should inherit from the parent. But whenever
downloading executable content, it should always be https.