Added the simple coil example image to the tutorial.
This example is a continuation of the tutorial up to this point. It uses
the same molecule structure combined with a few simple BNGL rules to
build the coil structures inside MCell itself. The example is only
briefly mentioned here.

The coil example should be covered in more depth in a subsequent
tutorial page introducing BNGL syntax and MCellR. That would be the
third page in the Spatially Structured Molecule tutorial section.