Clarified the difference between broken bonds and incomplete references
"Breaking a bond" simply means disconnecting the two components that had
been bound. That's not an error and does not trigger the red highlight.
The red highlight appears when the bonds are inconsistent. Inconsistent
bonds are those where the two components do not reference each other.
The text used the term "broken" when it should have been "inconsistent".